[b-97] changes after review
