+++ src/main/java/com/greetingcard/web/security/SecurityApplicationConfig.java
@@ -4,7 +4,6 @@ import com.greetingcard.web.security.jwt.JwtTokenVerifierFilter;
 import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.builders.WebSecurity;
@@ -15,8 +14,7 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 
 @Configuration
 @EnableWebSecurity
-@Order(1000)
-public class ApplicationSecurityConfig extends WebSecurityConfigurerAdapter {
+public class SecurityApplicationConfig extends WebSecurityConfigurerAdapter {
 
     @Autowired
     private JwtTokenVerifierFilter jwtTokenVerifierFilter;
