+++ src/main/java/com/greetingcard/web/security/jwt/JwtTokenVerifierFilter.java
@@ -34,7 +34,7 @@ public class JwtTokenVerifierFilter extends OncePerRequestFilter {
     protected void doFilterInternal(HttpServletRequest request,
                                     HttpServletResponse response,
                                     FilterChain filterChain) throws ServletException, IOException {
-
+        log.info("JwtTokenVerifierFilter");
         String token = getTokenFromRequest(request);
         if (token != null && jwtProvider.validateToken(token)) {
             String userLogin = jwtProvider.getLoginFromToken(token);
