+++ src/test/java/com/greetingcard/web/controller/AuthControllerSystemTest.java
@@ -1,0 +1,105 @@
+package com.greetingcard.web.controller;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.github.database.rider.core.api.configuration.DBUnit;
+import com.github.database.rider.core.api.configuration.Orthography;
+import com.github.database.rider.core.api.dataset.DataSet;
+import com.github.database.rider.spring.api.DBRider;
+import com.greetingcard.RootApplicationContext;
+import com.greetingcard.dao.jdbc.TestConfiguration;
+import org.flywaydb.core.Flyway;
+import org.junit.jupiter.api.*;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.mock.web.MockHttpServletResponse;
+import org.springframework.test.context.junit.jupiter.web.SpringJUnitWebConfig;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@DBRider
+@DBUnit(caseInsensitiveStrategy = Orthography.LOWERCASE)
+@DataSet(value = {"languages.xml", "types.xml", "roles.xml", "statuses.xml", "users.xml", "cards.xml", "cardsUsers.xml",
+        "congratulations.xml", "links.xml", "forgot_password_hashes.xml", "verify_email_hashes.xml"},
+        executeStatementsBefore = "SELECT setval('users_user_id_seq', 10);",
+        cleanAfter = true)
+@SpringJUnitWebConfig(value = {TestConfiguration.class, RootApplicationContext.class})
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+public class AuthControllerSystemTest {
+    @Autowired
+    private WebApplicationContext context;
+
+    private MockMvc mockMvc;
+
+    @Autowired
+    private Flyway flyway;
+
+    @BeforeAll
+    void dbSetUp() {
+        flyway.migrate();
+    }
+
+    @BeforeEach
+    void setMockMvc() {
+        mockMvc = MockMvcBuilders.webAppContextSetup(context).build();
+    }
+
+
+    @Test
+    @DisplayName("Login user")
+    void testLoginIfUserExist() throws Exception {
+        //prepare
+        Map<String, String> userCredential = new HashMap<>();
+        userCredential.put("login", "user");
+        userCredential.put("password", "user");
+        ObjectMapper mapper = new ObjectMapper();
+        String json = mapper.writeValueAsString(userCredential);
+
+        //when
+        MockHttpServletResponse response = mockMvc.perform(post("/api/v1/auth")
+                .contentType(APPLICATION_JSON_VALUE)
+                .accept(APPLICATION_JSON_VALUE)
+                .content(json))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.login").value("user"))
+                .andExpect(jsonPath("$.userId").value("2"))
+                .andReturn().getResponse();
+
+        String headerValue = response.getHeader("Authorization");
+        assertNotNull(headerValue);
+        assertTrue(headerValue.startsWith("Bearer "));
+    }
+
+    @Test
+    @DisplayName("Login user if login didn't create")
+    void testLoginIfUserIsNotExist() throws Exception {
+        //prepare
+        Map<String, String> userCredential = new HashMap<>();
+        userCredential.put("login", "user_don't_create");
+        userCredential.put("password", "user");
+        ObjectMapper mapper = new ObjectMapper();
+        String json = mapper.writeValueAsString(userCredential);
+
+        //when
+        mockMvc.perform(post("/api/v1/auth")
+                .contentType(APPLICATION_JSON_VALUE)
+                .accept(APPLICATION_JSON_VALUE)
+                .content(json))
+                .andDo(print())
+                .andExpect(status().isUnauthorized())
+                .andExpect(jsonPath("$.message").value("Access denied. Please check your login and password"));
+    }
+
+
+}
